// office-assistant/src/app/model/questionnaire/questionnaire.model.ts
export class Questionnaire {
    Id?: number;
    UserName?: string;
    FirstName: string;
    LastName: string;
    MothersName: string;
    FathersName: string;
    DateOfBirth: Date;
    Street: string;
    HomeNumber: number;
    ApartmentNumber: number;
    Education: string;
    SchoolName: string;
    LastEmployer: string;
    PhoneNumber: string;
    PeselNumber: string;
}